requirements/base.txt
django-cbv==0.1.5
